all_of_type also fits better in safe.py
